ovn-ic: Introduces I+P engine for Transit Switch.
This patch introduces the incremental processing engine
for transit switches. This engine aims to speed-up the
processing time of transit switches-related tasks.

According to tests made, it was possible
prevent 98% of full recomputes on the transit switches,
and when handling incrementally it was observed an speed-up
of 90% in runtime execution (from 60ms to 6ms).